Use cloud_image to name the image to use.  Update VM XML.
The bash script expected Cmd to be set to the image name but we're going
to use cloud_image for that.

I updated the XML to use x86_64 as well as bridge networking and set the
image size right in the XML.  I also added the mac_address thing but
that needs to be plumbed in properly.

Signed-off-by: Ian Main <imain@redhat.com>