Fix up XML and we're back to latest condor
So I found another different bug with older condor, so we're now using
upstream 7.6 which has a bugfix in it to make this work.  As such we're
back to vm_disk.

Also update XML to use {NAME}, it has to be the name as assigned by
condor or it will not be able to track the instance.

Also add ip/mac address assignment as a temporary thing to show how it
will work.

Signed-off-by: Ian Main <imain@redhat.com>