Remove existing user_data
The existing user_data setup is not what we want and actually causes
instances to not start by setting port to an empty string, so I'm just
ripping it out entirely for now.  This should be replaced with an
ec2-like user_data which will insert scripts into the image itself.

Signed-off-by: Ian Main <imain@redhat.com>